Don't run if the link location doesn't exist.

+++ node/add-to-path.js
@@ -45,6 +45,10 @@ const addToPath = async () => {
             linkDest = ""; // TODO.
         }
 
+        if (!fs.existsSync(linkDest)) {
+            return
+        }
+
         await runSudoCommand(`ln -fs ${linkDest} ${getLinkPath()}`, options);
     }
 }
